Convert `DataReferenceButton` to TypeScript
